fix(core): ValueCountMismatch carries a Long count; Edit.plan documented as a semantic dry-run
`XLError.ValueCountMismatch.expected` was an `Int`, and `EditInterpreter.countMatch`
truncated the `Long` cell count through `.toInt`: a `put-values` over `A1:XFD1048576`
(2^34 cells) read "expected 0 values, got 1" with the hint "provide exactly 0 values".
The field is now a `Long` end to end (every construction site widens; the CLI's
`countMismatch` passes an `Int` that widens too), pinned by a test over the whole grid.

`Edit.plan` was documented as producing the planned rows "without the workbook" while
it is `applyAll(...).map(_.planned)`. The interpreter's validation phase does not
separate from application (sheet resolution, shift refusals, structural bounds and
name lookups all fail inside `applyOne`), so a static plan would either duplicate the
interpreter's bookkeeping or diverge from `applyAll`'s failures. The doc now says what
it is: a semantic dry-run that applies and discards, failing exactly where `applyAll`
fails; `Edit.validate` is the static check. A test pins the failure parity, including a
refusal `validate` cannot see.

The CHANGELOG bullet says 49 cases (what `Edit` and `EditSchema.all` have), describes
`plan` as above and records the `Long` widening.

Refs #582

+++ xl-core/src/com/tjclp/xl/error/XLError.scala
@@ -82,8 +82,12 @@ enum XLError derives CanEqual:
   /** Security error (ZIP bomb, formula injection, size limits) */
   case SecurityError(reason: String)
 
-  /** Number of supplied values mismatched expectation */
-  case ValueCountMismatch(expected: Int, actual: Int, context: String)
+  /**
+   * Number of supplied values mismatched expectation. `expected` is a `Long`: a range's cell count
+   * (`CellRange.cellCount`) exceeds `Int.MaxValue` from `A1:XFD131072` on, and the message must
+   * report it rather than its truncation.
+   */
+  case ValueCountMismatch(expected: Long, actual: Int, context: String)
 
   /** Unsupported type in batch put operation */
   case UnsupportedType(ref: String, typeName: String)
